pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["ryfe <denis.koptelev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
django = "^5.0.6"
djangorestframework = "^3.15.1"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
requests = "^2.32.3"
coverage = "^7.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
